refactor

+++ src/world/generator/WorldGenerator.h
@@ -23,7 +23,7 @@ public:
         const Content* content
     );
 
-	virtual void generate(voxel* voxels, int x, int z, int seed);
+	virtual void generate(voxel* voxels, int x, int z, uint64_t seed);
 
     inline static std::string DEFAULT = BUILTIN_CONTENT_NAMESPACE + ":default";
 };
